Removed Tiger support. It's just not going to happen at this time. Maybe in a 2.1 release or something.

+++ Classes/ShowListDelegate.m
@@ -184,7 +184,7 @@
         [alert addButtonWithTitle:quitButton];
         [alert addButtonWithTitle:cancelButton];
 
-        NSInteger answer = [alert runModal];
+        int answer = [alert runModal];
         [alert release];
         alert = nil;
